+++ src/oci/database/models/patch_history_entry.py
@@ -2,7 +2,7 @@
 # Copyright (c) 2016, 2017, Oracle and/or its affiliates. All rights reserved.
 
 
-from ...util import formatted_flat_dict
+from ...util import formatted_flat_dict, NONE_SENTINEL, value_allowed_none_or_none_sentinel  # noqa: F401
 from ...decorators import init_model_state_from_kwargs
 
 
@@ -101,7 +101,7 @@ class PatchHistoryEntry(object):
         :type: str
         """
         allowed_values = ["APPLY", "PRECHECK"]
-        if action not in allowed_values:
+        if not value_allowed_none_or_none_sentinel(action, allowed_values):
             action = 'UNKNOWN_ENUM_VALUE'
         self._action = action
 
@@ -181,7 +181,7 @@ class PatchHistoryEntry(object):
         :type: str
         """
         allowed_values = ["IN_PROGRESS", "SUCCEEDED", "FAILED"]
-        if lifecycle_state not in allowed_values:
+        if not value_allowed_none_or_none_sentinel(lifecycle_state, allowed_values):
             lifecycle_state = 'UNKNOWN_ENUM_VALUE'
         self._lifecycle_state = lifecycle_state